Working getNextFragment routine

+++ src/UI/font.cpp
@@ -34,8 +34,8 @@ void CFont::loadFromStream(std::istream & input) {
 		glyphs[i].t = static_cast<float>(buffer.y) / height;
 		glyphs[i].v = glyphs[i].t + texLineHeight;
 
-		glyphs[i].width = (float)buffer.width;
-		glyphs[i].height = (float)lineHeight;
+		glyphs[i].width = buffer.width;
+		glyphs[i].height = lineHeight;
 		//Font->Glyphs[i].Rect.originX = 0;
 		//Font->Glyphs[i].Rect.originY = (float)line_height;
